.env on server, firebase config extraction
