+++ .gitignore
@@ -1,13 +1,10 @@
 .idea/
 .vscode/
-common/firebase.env
-
 .DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
 
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+common/firebase.env
+.env
