+++ server/db/reset.ts
@@ -1,3 +1,5 @@
+import '../env';
+
 import {
   ActivityStatus,
   ActivityType,
