+++ server/env.ts
@@ -2,4 +2,4 @@ import { resolve } from 'path';
 
 import { config } from 'dotenv';
 
-config({ path: resolve(__dirname, '../../local.env') });
+config({ path: resolve(__dirname, '../.env') });
